backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
ultralytics==8.0.196
torch>=1.13.0
torchvision>=0.14.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
cloudinary==1.36.0
python-dotenv==1.0.0
requests==2.31.0
